Added async callbacks for getting further information
I'm going to try to only commit stuff that actually works from now on. :
)... aside from that, this adds callbacks that can be used to do stuff
like when someone tries to pick up an item in a room and there's more
than one that fits the description, you can ask them which one. I'm
going to add separate infrastructure for checking if something exists in
a room, etc.